requirements.txt
# Function dependencies, for example:
# package>=version
pillow
urllib3
torchvision